+++ ikodave/src/test/java/com/example/problems/Filters/FilterStatusAttemptedTest.java
@@ -1,4 +1,4 @@
-package com.example.submissions.Filters;
+package com.example.problems.Filters;
 
 import com.example.problems.Filters.FilterStatusAttempted;
 import com.example.problems.Filters.Parameters.Parameter;
